+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # UNRELEASED
+- Recipes::Master: configuration files for scheduler and controller-manager
 - SSL: scheduler and controller_manager should have their own keypairs
 - [CNI/plugins: 0.8.5](https://github.com/containernetworking/plugins/releases/tag/v0.8.5)
 - [CoreDNS: 1.6.7](https://coredns.io/2020/01/28/coredns-1.6.7-release/)